+++ soap_server/src/main/java/by/maribo/web_service/repository/impl/Rule.java
@@ -1,10 +1,15 @@
 package by.maribo.web_service.repository.impl;
 
 import by.maribo.web_service.entity.Entity;
+import by.maribo.web_service.repository.CurrentEntityExistInRepository;
+import by.maribo.web_service.repository.EntityNotExistInRepository;
 import by.maribo.web_service.repository.EntityRepository;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import static by.maribo.web_service.repository.Message.EXIST_IN_REPOSITORY;
+import static by.maribo.web_service.repository.Message.IS_NOT_EXIST_IN_REPOSITORY;
 
 public class Rule implements EntityRepository {
 	private static final Rule instance = new Rule();
@@ -33,24 +38,32 @@ public class Rule implements EntityRepository {
 	}
 
 	@Override
-	public void add(Entity entity) {
+	public void add(Entity entity) throws CurrentEntityExistInRepository {
+		if (repository.contains(entity)) {
+			throw new CurrentEntityExistInRepository(entity.toString() + " " + EXIST_IN_REPOSITORY);
+		}
 		entity.setId(id++);
 		repository.add(entity);
 	}
 
 	@Override
-	public void delete(Entity entity) {
+	public void delete(Entity entity) throws EntityNotExistInRepository {
+		if (!repository.contains(entity)) {
+			throw new EntityNotExistInRepository(entity.toString() + " " + IS_NOT_EXIST_IN_REPOSITORY);
+		}
 		repository.remove(entity);
 	}
 
 	@Override
-	public void modify(int id, Entity entity) {
+	public void modify(int id, Entity entity) throws EntityNotExistInRepository {
 		for (Entity currentEntity : repository) {
 			if (currentEntity.getId() == id) {
 				currentEntity.setName(entity.getName());
 				currentEntity.setDescription(entity.getDescription());
+				return;
 			}
 		}
+		throw new EntityNotExistInRepository(entity.toString() + " " + IS_NOT_EXIST_IN_REPOSITORY);
 	}
 
 	@Override
